exist_tag_test

+++ Jenkinsfile
@@ -1,28 +1,38 @@
 pipeline {
     agent { label 'master' }
+
     /*
+    parameters {
+        gitParameter name: 'TAG', 
+                     type: 'PT_TAG',
+                     defaultValue: 'master'
+    }
     parameters {
         gitParameter branchFilter: 'origin/(.*)', defaultValue: 'master', name: 'BRANCH', type: 'PT_BRANCH'
     }
-
-export NVM_DIR="$HOME/.nvm"
-[ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"  # This loads nvm
-[ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"
     */
 
     environment {
         PROD_USER = 'deploy'
         PROD_HOST = '100.100.100.11'
         
+        /*
         //Имя тэга у последнего протегированного коммита
         NEW_TAG_NAME = sh(
             returnStdout: true,
             script: "git show-ref --tags |\
                      grep \$(git rev-list --tags --date-order | head -1) |\
                      awk -F / '{print \$NF}'"
-            ).trim()
-        
-        //Текущая версия релиза на production сервере 
+        ).trim()
+        */
+
+        // Проверка наличия дирректории с релизом на production сервере
+        TAG_EXIST_ON_PROD = sh(
+            returnStdout: true,
+            script: "ssh ${PROD_USER}@${PROD_HOST} \"if [[ -d /var/www/myapp/releases/${TAG_NAME} ]]; then echo true; else echo false; fi\""
+        ).trim()
+
+        // Текущая версия релиза на production сервере 
         CURRENT_VERSION_ON_PROD = sh(
             returnStdout: true,
             script: "ssh ${PROD_USER}@${PROD_HOST} 'cd /var/www/myapp; ls -la | grep current | cut -d \'/\' -f6' "
@@ -31,6 +41,9 @@ export NVM_DIR="$HOME/.nvm"
 
     stages{
         stage('Build') {
+            when {
+                expression { TAG_EXIST_ON_PROD == "false" }
+            }
             agent { label 'nodejs' }
             steps {
                     sh '''
@@ -51,7 +64,6 @@ export NVM_DIR="$HOME/.nvm"
                 }
             }
         }
-
         
         stage('Deploy') {
             agent { label 'deploy' }
@@ -61,23 +73,30 @@ export NVM_DIR="$HOME/.nvm"
                 - симлинк указывает на дирректорию имя которой отлично от последнего добавленного тэга
             */
             when {
-                 expression { CURRENT_VERSION_ON_PROD != NEW_TAG_NAME }
+                 expression { CURRENT_VERSION_ON_PROD != TAG_NAME }
             }
             steps {
-                sh 'git clean -fdx'
-                copyArtifacts filter: 'build.zip', fingerprintArtifacts: true, projectName: '${JOB_NAME}', selector: specific('${BUILD_NUMBER}')
-                unzip zipFile: 'build.zip', dir: './build'
-                sh '''
-                echo "CURRENT_VERSION_ON_PROD: ${CURRENT_VERSION_ON_PROD}"
-                echo "NEW_TAG_NAME: ${NEW_TAG_NAME}"
+                script {
+                    if (TAG_EXIST_ON_PROD == "true") {
+                        ssh ${PROD_USER}@${PROD_HOST} "ln -sfn /var/www/myapp/releases/${TAG_NAME}/ /var/www/myapp/current"
+                        echo "symlink changed to ${TAG_NAME}"
+                    } else {
+                        sh 'git clean -fdx'
+                        copyArtifacts filter: 'build.zip', fingerprintArtifacts: true, projectName: '${JOB_NAME}', selector: specific('${BUILD_NUMBER}')
+                        unzip zipFile: 'build.zip', dir: './build'
+                        sh '''
+                        echo "CURRENT_VERSION_ON_PROD: ${CURRENT_VERSION_ON_PROD}"
+                        echo "TAG_NAME: ${TAG_NAME}"
 
-                rsync -avr ./build/ ${PROD_USER}@${PROD_HOST}:/var/www/myapp/releases/${NEW_TAG_NAME}/
-                
-                ssh ${PROD_USER}@${PROD_HOST} "ln -sfn /var/www/myapp/releases/${NEW_TAG_NAME}/ /var/www/myapp/current"
-                '''
-                sh '''
-                ssh ${PROD_USER}@${PROD_HOST} 'cd /var/www/myapp/releases && rm -rf $(ls -t | awk "NR>5")'
-                '''        
+                        rsync -avr ./build/ ${PROD_USER}@${PROD_HOST}:/var/www/myapp/releases/${TAG_NAME}/
+                        
+                        ssh ${PROD_USER}@${PROD_HOST} "ln -sfn /var/www/myapp/releases/${TAG_NAME}/ /var/www/myapp/current"
+                        '''
+                        sh '''
+                        ssh ${PROD_USER}@${PROD_HOST} 'cd /var/www/myapp/releases && rm -rf $(ls -t | awk "NR>5")'
+                        ''' 
+                    }
+                }
             }
         }     
     }
